requirements.txt
web3
eth_abi
pydantic
colorlog~=5.0.1
python-dotenv~=1.0.0